+++ test/integration/integration_test.js
@@ -35,7 +35,7 @@ suite('integration tests', function() {
   suite('init templates', () => {
 
     // TODO(#562): enable test commands.
-    test('test the polymer 1.x application template', () => {
+    test('test the Polymer 1.x application template', () => {
       let dir;
       return runGenerator(createApplicationGenerator('polymer-1.x'))
         .withPrompts({ name: 'my-app' }) // Mock the prompt answers
@@ -47,7 +47,7 @@ suite('integration tests', function() {
         .then(() => runCommand(binPath, ['build'], {cwd: dir}));
     });
 
-    test('test the polymer 2.x application template', () => {
+    test('test the Polymer 2.x application template', () => {
       let dir;
       return runGenerator(createApplicationGenerator('polymer-2.x'))
         .withPrompts({ name: 'my-app' }) // Mock the prompt answers
@@ -59,7 +59,7 @@ suite('integration tests', function() {
         .then(() => runCommand(binPath, ['build'], {cwd: dir}));
     });
 
-    test('test the polymer 2.x "element" template', () => {
+    test('test the Polymer 2.x "element" template', () => {
       let dir;
       return runGenerator(createElementGenerator('polymer-2.x'))
         .withPrompts({ name: 'my-element' }) // Mock the prompt answers
@@ -70,7 +70,7 @@ suite('integration tests', function() {
         // .then(() => runCommand(binPath, ['test'], {cwd: dir}));
     });
 
-    test('test the polymer 1.x "element" template', () => {
+    test('test the Polymer 1.x "element" template', () => {
       let dir;
       return runGenerator(createElementGenerator('polymer-1.x'))
         .withPrompts({ name: 'my-element' }) // Mock the prompt answers
@@ -103,11 +103,12 @@ suite('integration tests', function() {
     // TODO(justinfagnani): consider removing these integration tests
     // or checking in the contents so that we're not subject to the
     // other repo changing
-    test.skip('test the "starter-kit" template', () => {
+    test.skip('test the Polymer 1.x "starter-kit" template', () => {
       let dir;
       const PSKGenerator = createGithubGenerator({
         owner: 'PolymerElements',
         repo: 'polymer-starter-kit',
+        semverRange: '^2.0.0',
         githubToken,
       });
 
@@ -121,6 +122,30 @@ suite('integration tests', function() {
         // .then(() => runCommand(binPath, ['test'], {cwd: dir}))
         .then(() => runCommand(binPath, ['build'], {cwd: dir}));
     });
+
+    // TODO(justinfagnani): consider removing these integration tests
+    // or checking in the contents so that we're not subject to the
+    // other repo changing
+    test.skip('test the Polymer 2.x "starter-kit" template', () => {
+      let dir;
+      const PSKGenerator = createGithubGenerator({
+        owner: 'PolymerElements',
+        repo: 'polymer-starter-kit',
+        semverRange: '^3.0.0',
+        githubToken,
+      });
+
+      return runGenerator(PSKGenerator)
+        .toPromise()
+        .then((_dir) => { dir = _dir })
+        .then(() => runCommand(binPath, ['install'], { cwd: dir }))
+        // .then(() => runCommand(
+        //     binPath, ['experimental-lint', '--rules=polymer-2'],
+        //     { cwd: dir }))
+        // .then(() => runCommand(binPath, ['test'], {cwd: dir}))
+        .then(() => runCommand(binPath, ['build'], {cwd: dir}));
+    });
+
 
   });