+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Upcoming
 
+- Add PSK 3.0 (Polymer 2.0 Polymer Starter Kit) template to the init command.
+
 ## v0.18.0-pre.10 [02-21-2017]
 
 - **New `build` Behavior**: New build options have been added to give you more control over the generated build. These options can be defined in your project's `polymer.json`, or via CLI flags. Run `polymer build --help` to see a list of new supported CLI flags.